#20 Moved INode to abstractions
