+++ QuickMaths.General/Abstractions/INode.cs
@@ -1,0 +1,12 @@
+﻿namespace QuickMaths.General.Abstractions;
+
+internal interface INode
+{
+    public double Calculate();
+
+    public IArithmeticable GetDerivative();
+
+    public void SetVariables(Dictionary<string, double> variables);
+
+    public int GetPriority();
+}
